fix: 64-bit widget ids on Android — buttons/events now reach the VM
Root cause, proven by the on-device overlay: registered callback keys were
`4294967304:press` (= 2^32+8) while fired events were `9:press`. The VM manager
packs the owning VM id into the high 32 bits of every widget handle
(encode_handle => 2^32*vm + local), so ids exceed 32-bit range. The Android
controller parsed them with Kotlin `Int`/JSON `getInt` and JNI `jint`, which
truncated 2^32+9 -> 9. The native side was internally consistent at the
truncated value (create/connect/fire all truncated the same way), but it pushed
`9` back while the JS dispatcher held the full `4294967305`, so every callback
lookup missed and no handler ever ran (inv stayed 0).

Carry widget ids as 64-bit end to end on the native side:
- WidgetController.kt: entries keyed by Long; getLong for id/p/c/cb; all id
  params Long.
- VictorSurfaceView.kt + JNI nativePushEvent: jlong id.
- ElpianWidgetsJsi.cpp: push_event / ElpianWidgetsPushEvent take long long.
- iOS ElpianWidgetsBridge: pushEvent id as long long (Swift Int is already 64-bit).

JS never truncates (Number is 64-bit-safe to 2^53), so no JS change. Added a
dispatcher regression test: a 2^32+9 handle connects and fires exactly, and the
truncated id 9 must miss.

+++ victor/react-native/modules/elpian-widgets/android/src/main/cpp/ElpianWidgetsJsi.cpp
@@ -104,7 +104,7 @@ std::string drain_ops(const std::string &app) {
   return out;
 }
 
-void push_event(const std::string &app, int id, const std::string &event,
+void push_event(const std::string &app, long long id, const std::string &event,
                 const char *argJson) {
   std::string elem = "[" + std::to_string(id) + ",\"" + event + "\"," +
                      std::string(argJson && argJson[0] ? argJson : "null") + "]";
@@ -134,7 +134,7 @@ Java_expo_modules_elpianwidgets_VictorSurfaceView_nativePollOps(JNIEnv *env, job
 
 extern "C" JNIEXPORT void JNICALL
 Java_expo_modules_elpianwidgets_VictorSurfaceView_nativePushEvent(
-    JNIEnv *env, jobject, jstring appId, jint id, jstring event, jstring argJson) {
+    JNIEnv *env, jobject, jstring appId, jlong id, jstring event, jstring argJson) {
   const char *app = env->GetStringUTFChars(appId, nullptr);
   const char *e = env->GetStringUTFChars(event, nullptr);
   const char *a = argJson ? env->GetStringUTFChars(argJson, nullptr) : nullptr;
@@ -161,7 +161,7 @@ extern "C" const char *ElpianWidgetsDrainOps(const char *appId) {
   return buf;
 }
 
-extern "C" void ElpianWidgetsPushEvent(const char *appId, int id, const char *event,
+extern "C" void ElpianWidgetsPushEvent(const char *appId, long long id, const char *event,
                                        const char *argJson) {
   elpianwidgets::push_event(appId ? appId : "main", id, event ? event : "", argJson);
 }
